temporarily hard code mermaid script
Store isn't compatible with static rendering of pages? I think the
markup rendering of mermaid diagrams will have to be transitioned to a
shortcode if I want it to be automagically included.

I know including this outside <head> isn't really proper, but I don't
want to include it where it isn't needed.